requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

pandas==2.1.1
Pillow==9.4.0
streamlit==1.32.0
